+++ hf2p_server/source/networking/messages/network_message_queue.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <windows.h>
+#include "..\..\cseries\cseries.h"
 #include "..\..\memory\sliding_window.h"
 #include "network_message_type_collection.h"
 #include "..\delivery\network_connection.h"
@@ -14,8 +14,8 @@ class c_network_message_queue : c_network_channel_client
 		{
 			byte flags;
 			byte fragment_size_bytes;
-			uint16_t fragment_size_bits;
-			int32_t packet_sequence_number;
+			ushort fragment_size_bits;
+			long packet_sequence_number;
 			s_outgoing_fragment_record* fragment;
 		};
 
@@ -23,7 +23,7 @@ class c_network_message_queue : c_network_channel_client
 		{
 			byte flags;
 			byte fragment_size_bytes;
-			uint16_t fragment_size_bits;
+			ushort fragment_size_bits;
 			s_incoming_fragment_record* fragment;
 			byte __data[32];
 		};
@@ -35,7 +35,7 @@ class c_network_message_queue : c_network_channel_client
 		s_outgoing_fragment_record* m_outgoing_fragments[2];
 		c_sliding_window m_incoming_window;
 		s_incoming_fragment_record* m_incoming_fragments[2];
-		int32_t m_outgoing_payload_bytes;
-		int32_t m_incoming_payload_bytes;
+		long m_outgoing_payload_bytes;
+		long m_incoming_payload_bytes;
 };
 static_assert(sizeof(c_network_message_queue) == 0x64);
